requirements.txt
pymilvus
sentence-transformers
openai
python-dotenv
numpy
pandas
pytest
fastapi
uvicorn
qdrant-client 